Ported SCSS Layout css to new Theme. Refactor of themes folder.
