+++ packages/ui-lib/src/theme/index.ts
@@ -1,6 +1,6 @@
 import { fontFamily, dimensions, animation, deviceSize, deviceMediaQuery } from './common';
-import { typography } from './typography';
-import { colors } from './colors';
-import { styles } from './styles';
-import { custom } from './custom';
+import { typography } from './legacy/typography';
+import { colors } from './legacy/colors';
+import { styles } from './legacy/styles';
+import { custom } from './legacy/custom';
 export default { fontFamily, dimensions, typography, colors, styles, animation, custom, deviceSize, deviceMediaQuery };
