+++ packages/ui-lib/src/theme/variables/Layout.ts
@@ -1,0 +1,52 @@
+import { css } from 'styled-components';
+import { deviceMediaQuery } from '../common';
+
+export const layoutVariables = css`
+  /** Layouts */
+  :root {
+    --max-content-width: 1840px;
+  }
+
+  .content-layout {
+    
+    &-default {
+      --content-layout-padding-top: 32px;
+      --content-layout-padding-right: 48px;
+      --content-layout-padding-bottom: 64px;
+      --content-layout-padding-left: 64px;
+      --utility-header-padding-top: 48px;
+
+      @media ${deviceMediaQuery.xlarge}{
+        --content-layout-padding-top: 32px;
+        --content-layout-padding-right: 96px;
+        --content-layout-padding-bottom: 64px;
+        --content-layout-padding-left: 96px;
+        --utility-header-padding-top: 48px;
+      }
+
+      @media ${deviceMediaQuery.xxlarge}{
+        --content-layout-padding-top: 30px;
+        --content-layout-padding-right: 120px;
+        --content-layout-padding-bottom: 64px;
+        --content-layout-padding-left: 120px;
+        --utility-header-padding-top: 48px;
+      }
+    }
+
+    &-fullscreen {
+      --content-layout-padding-top: 0px;
+      --content-layout-padding-right: 0px;
+      --content-layout-padding-bottom: 0px;
+      --content-layout-padding-left: 0px;
+      --utility-header-padding-top: 0px;
+    }
+
+    &-dashboard {
+      --content-layout-padding-top: 32px;
+      --content-layout-padding-right: 32px;
+      --content-layout-padding-bottom: 32px;
+      --content-layout-padding-left: 32px;
+      --utility-header-padding-top: 32px;
+    }
+  }
+`
